refactor patientResource templates
